+++ tests/playlist.rs
@@ -20,16 +20,16 @@ fn test_mp3_url() {
 }
 
 #[test]
-fn test_mpeg_dash_manifest_1() {
+fn test_mpeg_dash_mpd_1() {
     let fmt = FileFormat::from_file("fixtures/playlist/sample1.mpd").unwrap();
-    assert_eq!(fmt, FileFormat::MpegDashManifest);
+    assert_eq!(fmt, FileFormat::MpegDashMpd);
 }
 
 #[cfg(feature = "reader-xml")]
 #[test]
-fn test_mpeg_dash_manifest_2() {
+fn test_mpeg_dash_mpd_2() {
     let fmt = FileFormat::from_file("fixtures/playlist/sample2.mpd").unwrap();
-    assert_eq!(fmt, FileFormat::MpegDashManifest);
+    assert_eq!(fmt, FileFormat::MpegDashMpd);
 }
 
 #[test]